refactor(session): extract cvg session suite into convergio-cli-session crate
`convergio-cli` shipped at 11033 lines after PR #164 — past the
11000-line per-crate hard cap (CONSTITUTION § 13). PR #166
(SessionStart hook + telemetry) closed because adding ~500 lines
would have pushed the crate further over.

Mirror ADR-0040's coherence-extraction fix: move the `cvg session`
suite (resume + pre-stop + checks/) into a new cross-cutting
`convergio-cli-session` crate. Pure refactor — zero behaviour
change. The CLI keeps a 35-line shim at `commands/session.rs` so
the `main.rs` dispatcher line is unchanged.

Drop in convergio-cli: 11033 -> 9138 (-1379 lines after auto-doc
regen). New crate: 1499 lines.

Tracks: 07418af1-4aef-4cc8-8439-e3693c3d73bc
