+++ crates/convergio-cli-session/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "convergio-cli-session"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Session lifecycle commands (cvg session) extracted from convergio-cli"
+readme = "README.md"
+keywords = ["session", "convergio", "cli"]
+categories = ["development-tools"]
+
+[lints]
+workspace = true
+
+[dependencies]
+convergio-i18n = { workspace = true }
+anyhow = { workspace = true }
+chrono = { workspace = true }
+clap = { workspace = true }
+reqwest = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+tokio = { workspace = true }
+tracing = { workspace = true }
