+++ crates/convergio-cli-session/src/session_tests.rs
