+++ crates/convergio-cli/src/commands/mod.rs
@@ -39,10 +39,6 @@ mod pr_sync;
 mod pr_sync_parse;
 pub mod service;
 pub mod session;
-mod session_checks;
-mod session_pre_stop;
-mod session_pre_stop_run;
-mod session_render;
 pub mod setup;
 mod setup_repo_path;
 pub mod solve;
